services/orchestrator/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
asyncpg==0.29.0
redis==5.0.1
aiohttp==3.9.1
python-multipart==0.0.6
pyyaml==6.0.1
pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.3
httpx==0.26.0
aiortc==1.6.0
minio==7.2.3
celery==5.3.6
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
